leveleditor: A couple new feautres
Replaced the --clash launch option with --server. This will pave the way for future updates to the level editor that are tailored for specific Toontown games. Right now, there are four modes: "online", "rewritten", "clash", and "offline". Only Clash has a difference at this time.

Added a custom hoods feature. Look at the sample.json file in the hoods folder for a demo on how to create a custom hood. Custom hoods will be loaded automatically when their .json file is present in the hoods folder.