fps: Improve FPS meter
Had to do things a bit hacky here. Overwrote the normal function that initiates the FPS meter just so it doesn't get used normally. I tried to initialize the new FPS meter under it, but it won't work since it runs so early that you can't even load fonts yet. Thus, I made a new function for it.